refactor: rename FilterTreeView to FilterTreeDataProvider
Rename `FilterTreeView.ts` to `FilterTreeDataProvider.ts` to better
reflect its role as a data provider implementation.

This naming convention aligns with VS Code extension API patterns
and improves code readability.
